Set up to implement banned countries configuration
